Update container files to use systemd

+++ contrib/vm/container_manager.py
@@ -68,10 +68,11 @@ imagemagick
 iptables
 libcairo-dev
 libffi-dev
+libsystemd-dev
 m4
 nginx
 npm
-php5-fpm
+php7.0-fpm
 rsyslog-relp
 texlive
 texlive-latex-extra
@@ -98,9 +99,7 @@ sage_optional_packages = [
 "4ti2",
 "biopython",
 "cbc",
-"chomp",
 "cluster_seed",
-"coxeter3",
 "cryptominisat",
 "database_cremona_ellcurve",
 "database_gap",
@@ -138,6 +137,7 @@ python_packages = [
 "paramiko",
 "psutil",
 "sockjs-tornado",
+"git+https://github.com/systemd/python-systemd.git",
 # Optional
 "bitarray",
 "ggplot",
@@ -198,33 +198,46 @@ if $syslogfacility-text == "local3" then
 
 # HA-Proxy configuration is regenerated every time the script is run.
 HAProxy_header = """\
+# Default from Ubuntu 16.04 LTS
 global
-    chroot /var/lib/haproxy
-    daemon
-    group haproxy
-    user haproxy
-    log /dev/log local0
-    stats socket /var/run/haproxy.sock mode 600
+        log /dev/log    local0
+        log /dev/log    local1 notice
+        chroot /var/lib/haproxy
+        stats socket /run/haproxy/admin.sock mode 660 level admin
+        stats timeout 30s
+        user haproxy
+        group haproxy
+        daemon
 
+        # Default SSL material locations
+        ca-base /etc/ssl/certs
+        crt-base /etc/ssl/private
+
+        # Default ciphers to use on SSL-enabled listening sockets.
+        # For more information, see ciphers(1SSL). This list is from:
+        #  https://hynek.me/articles/hardening-your-web-servers-ssl-ciphers/
+        ssl-default-bind-ciphers ECDH+AESGCM:DH+AESGCM:ECDH+AES256:DH+AES256:ECDH+AES128:DH+AES:ECDH+3DES:DH+3DES:RSA+AESGCM:RSA+AES:RSA+3DES:!aNULL:!MD5:!DSS
+        ssl-default-bind-options no-sslv3
 
 defaults
-    log global
-    mode http
-    option dontlognull
-    option http-server-close
-    option httplog
-    option redispatch
-    timeout connect 5s
-    timeout client 50s
-    timeout server 50s
+        log     global
+        mode    http
+        option  httplog
+        option  dontlognull
+        timeout connect 5000
+        timeout client  50000
+        timeout server  50000
+        errorfile 400 /etc/haproxy/errors/400.http
+        errorfile 403 /etc/haproxy/errors/403.http
+        errorfile 408 /etc/haproxy/errors/408.http
+        errorfile 500 /etc/haproxy/errors/500.http
+        errorfile 502 /etc/haproxy/errors/502.http
+        errorfile 503 /etc/haproxy/errors/503.http
+        errorfile 504 /etc/haproxy/errors/504.http
 
-    errorfile 400 /etc/haproxy/errors/400.http
-    errorfile 403 /etc/haproxy/errors/403.http
-    errorfile 408 /etc/haproxy/errors/408.http
-    errorfile 500 /etc/haproxy/errors/500.http
-    errorfile 502 /etc/haproxy/errors/502.http
-    errorfile 503 /etc/haproxy/errors/503.http
-    errorfile 504 /etc/haproxy/errors/504.http
+# SageMathCell additions
+        option http-server-close
+        option redispatch
 """
 
 # {suffix} {port} {hostname} {peer_port} have to be set once
@@ -292,18 +305,6 @@ def communicate(command, message):
             msg = "{} failed".format(command)
             log.error(msg)
             raise RuntimeError(msg)
-
-
-def remove_pattern(path, pattern):
-    r"""
-    Remove all files and directories at ``path`` matching ``pattern``.
-    """
-    for name in fnmatch.filter(os.listdir(path), pattern):
-        full = os.path.join(path, name)
-        if os.path.isdir(full):
-            shutil.rmtree(full)
-        else:
-            os.remove(full)
 
 
 def timer_delay(delay, test=None):
@@ -528,6 +529,7 @@ def install_config_files():
         for file in files:
             name = os.path.join(root, file)
             adjust_names(shutil.copy(name, name[1:]))
+    check_call("systemctl enable sagecell")
     with open("/etc/network/interfaces", "a") as f:
         f.write("    up /root/firewall\n")
 
@@ -562,10 +564,10 @@ class SCLXC(object):
 
     def clone(self, clone_name, autostart=False, update=False):
         r"""
-        Create self, destroy old clone, and create it again.
+        Clone self, create a base container and destroy old clone if necessary.
         """
-        if not self.c.defined:
-            raise RuntimeError("cannot clone a non-existing container")
+        if not self.is_defined():
+            self.create()
         if update:
             self.update()
         self.shutdown()
@@ -586,30 +588,15 @@ class SCLXC(object):
 
     def create(self):
         r"""
-        Destroy and recreate self.
+        Create a base contrainer, destroy old one if necessary.
         """
         self.destroy()
         log.info("creating %s", self.name)
         # Try to automatically pick up proxy from host
         os.environ["HTTP_PROXY"] = "apt"
-
-        # if not self.c.create("ubuntu", 0, {"packages": ",".join(packages)}):
-        #     raise RuntimeError("failed to create " + self.name)
-        # Try to work around https://github.com/lxc/lxc/issues/283
-        cmd = "lxc-create -n {} -t ubuntu -B btrfs -- --packages={}"
-        check_call(cmd.format(self.name, ",".join(packages)))
-        self.c = lxc.Container(self.name)
-
-        # Try to work around https://github.com/lxc/lxc/issues/280
-        cmd = "apt-config shell APT_PROXY Acquire::http::Proxy"
-        try:
-            APT_PROXY = check_output(cmd).split("'")[1]
-            proxy_file = "/var/cache/lxc/trusty/rootfs-amd64" \
-                "/etc/apt/apt.conf.d/70proxy"
-            with open(proxy_file, "w") as f:
-                f.write(r'Acquire::http::Proxy "{}";'.format(APT_PROXY))
-        except IndexError:
-            pass
+        if not self.c.create(
+            "ubuntu", 0, {"packages": ",".join(packages)}, "btrfs"):
+                raise RuntimeError("failed to create " + self.name)
         os.environ.unsetenv("HTTP_PROXY")
 
         self.inside("/usr/sbin/deluser ubuntu --remove-home")
@@ -618,6 +605,7 @@ class SCLXC(object):
         self.inside(os.symlink, "/usr/bin/nodejs", "/usr/bin/node")
         log.info("installing npm packages")
         self.inside("npm install -g inherits requirejs coffee-script")
+        self.update()
 
     def destroy(self):
         r"""
@@ -747,6 +735,7 @@ class SCLXC(object):
         copyname = "container_logs/%s to %s on %s" % (start, end, self.name)
         if not os.path.exists("container_logs"):
             os.mkdir("container_logs")
+        log.debug("saving %s", copyname)
         shutil.copy(logname, copyname)
         check_call("bzip2 '{}'".format(copyname))
 
@@ -767,20 +756,18 @@ class SCLXC(object):
         r"""
         Update OS packages in ``self``.
         """
-        log.info("updating packages in %s", self.name)
-        self.inside("apt-get update")
-        self.inside("apt-get dist-upgrade -y --auto-remove")
+        if self.is_defined():
+            log.info("updating packages in %s", self.name)
+            self.inside("apt-get update")
+            self.inside("apt-get dist-upgrade -y --auto-remove")
+        else:
+            self.create()
 
 
 def restart_haproxy(names, backup_names=[]):
     r"""
     Regenerate HA-Proxy configuration file and restart it.
     """
-    # Make sure we have a fresh enough HA-Proxy
-    if check_output("haproxy -v").startswith("HA-Proxy version 1.4"):
-        log.info("HAProxy is too old, installing from backports")
-        check_call("apt-get install --target-release trusty-backports haproxy")
-
     log.debug("generating HAProxy configuration file")
     lines = [HAProxy_header]
     if names:
@@ -800,7 +787,8 @@ def restart_haproxy(names, backup_names=[]):
                 l += " backup"
                 for i, n in enumerate(backup_names):
                     lines.append(l.format(node=n, id=i + shift(n)))
-    if SCLXC(lxcn_tester).is_defined():
+    tester = SCLXC(lxcn_tester)
+    if tester.is_defined():
         section = HAProxy_section
         for k, v in {"port" : 8888,
                      "suffix": "_test",
@@ -817,10 +805,13 @@ def restart_haproxy(names, backup_names=[]):
     # names are not resolvable until they have started.
     with open("/etc/cron.d/haproxy", "w") as f:
         delay = start_delay * (len(up_names) + 2)
-        f.write("@reboot root sleep %d; service haproxy start\n" % delay)
+        f.write("@reboot root sleep %d; systemctl start haproxy\n" % delay)
     # Make it possible to use container names
     try:
-        check_call("host {}.lxc".format(lxcn_sagecell))
+        if names:
+            check_call("host {}.lxc".format(names[0]))
+        elif tester.is_defined():
+            check_call("host {}.lxc".format(lxcn_tester))
     except subprocess.CalledProcessError:
         log.info("making container names resolvable to IP addresses")
         with open("/etc/default/lxc-net", "r") as f:
@@ -832,7 +823,10 @@ def restart_haproxy(names, backup_names=[]):
             f.write("server=/lxc/10.0.3.1\ninterface=lo\n")
         log.info("Reboot for system configuration changes to take effect.")
         exit()
-    check_call("service haproxy reload")
+    try:
+        check_call("systemctl reload haproxy")
+    except subprocess.CalledProcessError:
+        check_call("systemctl start haproxy")
 
 
 logging.config.dictConfig(yaml.load("""
@@ -901,11 +895,12 @@ if not os.path.exists("/etc/rsyslog.d/sagecell.conf"):
     log.info("setting up rsyslog configuration file")
     with open("/etc/rsyslog.d/sagecell.conf", "w") as f:
         f.write(rsyslog_conf)
-    check_call("service rsyslog restart")
+    check_call("systemctl restart rsyslog")
 
 # Main chain: base -- precell -- (sagecell, backup)
+base = SCLXC(lxcn_base)
 if args.base:
-    SCLXC(lxcn_base).create()
+    base.create()
 
 sagecell = SCLXC(lxcn_sagecell)
 if args.savemaster:
@@ -921,11 +916,6 @@ if args.master or not sagecell.is_defined():
             precell.inside(
                 "su -c 'git -C /home/{server}/github/sagecell pull' {server}")
     else:
-        base = SCLXC(lxcn_base)
-        if base.is_defined():
-            base.update()
-        else:
-            base.create()
         precell = base.clone(lxcn_precell)
         precell.prepare_for_sagecell(args.keeprepos)
     sagecell = precell.clone(lxcn_sagecell)
@@ -964,9 +954,6 @@ if args.deploy:
         restart_haproxy(up_names, old_names)
         log.info("waiting for users to stop working with old containers...")
         timer_delay(deploy_delay, test)
-        # Make sure sagecell has an associated IP for restart_haproxy
-        sagecell.start()
-        sagecell.shutdown()
 
 restart_haproxy(up_names)
